Restructured project as discussed in the standup
