--- src/App.css
@@ -1,22 +1,0 @@
-.container {
-    display: flex;
-    flex-direction: column;
-    align-items: center;
-    justify-content: center;
-    width: 60%;
-    margin: auto;
-    padding: 1.5rem;
-}
-
-.title {
-    text-align: center;
-    font-size: 35px ;
-}
-.btn-container{
-    padding: 1rem;
-    display: flex;
-    flex-direction: row;
-    align-items: center;
-    justify-content: space-evenly;
-    width: 25%;
-}
